requirements.txt
torch>=2.0.0
numpy>=1.24.0
pygame>=2.1.0
Pillow>=9.0.0
